Reworked consumptions

+++ AppNiZiAPI/Models/Repositories/ConsumptionRespository.cs
@@ -8,8 +8,8 @@ namespace AppNiZiAPI
 {
     public class ConsumptionRespository : Repository, IConsumptionRepository
     {
-        
-        public bool AddConsumption(ConsumptionView consumption, int patientId)
+
+        public bool AddConsumption(Consumption consumption)
         {
             bool added;
             var insert = $"INSERT INTO Consumption " +
@@ -19,7 +19,7 @@ namespace AppNiZiAPI
             conn.Open();
             try
             {
-                added = ConsumptionCommand(insertQuery, consumption, patientId).ExecuteNonQuery() > 0;
+                added = ConsumptionCommand1(insertQuery, consumption).ExecuteNonQuery() > 0;
             }
             catch (Exception)
             {
@@ -29,35 +29,10 @@ namespace AppNiZiAPI
             return added;
         }
 
-        public int GetConsumptionPatientId(int consumptionId)
+        public bool DeleteConsumption(int consumptionId, int patientId)
         {
-            //TODO: Test method
+            if (patientId == 0) return false;
 
-            int patientId = 0;
-            var query = $"SELECT Consumption.patient_id FROM Consumption WHERE id = '{consumptionId}'";
-            using (SqlCommand sqlCommand = new SqlCommand(query, conn))
-            {
-                conn.Open();
-                try
-                {
-                    SqlDataReader reader = sqlCommand.ExecuteReader();
-                    while (reader.Read())
-                    {
-                        patientId = (int)reader["patient_id"];
-                    }
-                }
-                catch (Exception e)
-                {
-                    // logging?
-                    throw e;
-                }
-                conn.Close();
-            }
-            return patientId;
-        }
-
-        private bool DeleteConsumption(int consumptionId)
-        {
             bool affected;
             var query = $"DELETE FROM Consumption WHERE id = '{consumptionId}'";
             conn.Open();
@@ -74,19 +49,15 @@ namespace AppNiZiAPI
             return affected;
         }
 
-        public bool DeleteConsumption(int consumptionId, int patientId)
+        public ConsumptionView GetConsumptionByConsumptionId(int consumptionId)
         {
-            if (patientId == 0) return false;
-            if (GetConsumptionPatientId(consumptionId) ==  patientId) return DeleteConsumption(consumptionId);
-            return false;
-        }
-
-        public Consumption GetConsumptionByConsumptionId(int consumptionId)
-        {
-            var query = $"SELECT * FROM Consumption WHERE id = '{consumptionId}'";
+            var query = $"SELECT Consumption.*, WeightUnit.short, WeightUnit.unit " +
+                $"FROM Consumption " +
+                $"INNER JOIN WeightUnit ON Consumption.weight_unit_id = WeightUnit.id " +
+                $"WHERE Consumption.id = '{consumptionId}'";
 
             conn.Open();
-            Consumption consumption = new Consumption();
+            ConsumptionView consumption = new ConsumptionView();
             using (SqlCommand sqlCommand = new SqlCommand(query, conn))
             {
                 try
@@ -94,7 +65,7 @@ namespace AppNiZiAPI
                     SqlDataReader reader = sqlCommand.ExecuteReader();
                     while (reader.Read())
                     {
-                        consumption.Id = (int)reader["id"];
+                        consumption.ConsumptionId = (int)reader["id"];
                         consumption.FoodName = reader["food_name"].ToString();
                         consumption.KCal = (float)Convert.ToDouble(reader["kcal"]);
                         consumption.Protein = (float)Convert.ToDouble(reader["protein"]);
@@ -102,30 +73,41 @@ namespace AppNiZiAPI
                         consumption.Calium = (float)Convert.ToDouble(reader["calium"]);
                         consumption.Sodium = (float)Convert.ToDouble(reader["sodium"]);
                         consumption.Amount = (int)reader["amount"];
-                        consumption.WeightUnitId = (int)reader["weight_unit_id"];
+                        consumption.Weight = new WeightUnitModel
+                        {
+                            Id = (int)reader["weight_unit_id"],
+                            Short = (string)reader["short"],
+                            Unit = (string)reader["unit"]
+                        };
                         consumption.Date = Convert.ToDateTime(reader["date"]).Date;
                         consumption.PatientId = (int)reader["patient_id"];
+                        consumption.Valid = true;
                     }
                 }
                 catch (Exception e)
                 {
-                    // logging?
-                    throw e;
+                    consumption.Valid = false;
+                    return consumption;
                 }
             }
             conn.Close();
             return consumption;
         }
 
-        public List<ConsumptionView> GetConsumptionsForPatientBetweenDates(int patientId, DateTime startDate, DateTime endDate)
+        public List<PatientConsumptionView> GetConsumptionsForPatientBetweenDates(int patientId, DateTime startDate, DateTime endDate)
         {
-            String sqlStartDate = startDate.Date.ToString("yyyy-MM-dd").Replace("/", "-");
-            String sqlEndDate = endDate.Date.ToString("yyyy-MM-dd").Replace("/", "-");
-            var query = $"SELECT * FROM Consumption WHERE patient_id = '{patientId}' AND date BETWEEN '{sqlStartDate}' AND '{sqlEndDate}'";
+            string sqlStartDate = startDate.Date.ToString("yyyy-MM-dd").Replace("/", "-");
+            string sqlEndDate = endDate.Date.ToString("yyyy-MM-dd").Replace("/", "-");
+            
+            var query = $"SELECT Consumption.*, WeightUnit.short, WeightUnit.unit " +
+                $"FROM Consumption " +
+                $"INNER JOIN WeightUnit ON Consumption.weight_unit_id = WeightUnit.id " +
+                $"WHERE Consumption.patient_id = '{patientId}' AND Consumption.date BETWEEN '{sqlStartDate}' AND '{sqlEndDate}'";
 
-            List<ConsumptionView> consumptions = new List<ConsumptionView>();
+            List<PatientConsumptionView> consumptions = new List<PatientConsumptionView>();
             conn.Open();
 
+            PatientConsumptionView consumption;
             using (SqlCommand sqlCommand = new SqlCommand(query, conn))
             {
                 try
@@ -134,24 +116,41 @@ namespace AppNiZiAPI
                     SqlDataReader reader = sqlCommand.ExecuteReader();
                     while (reader.Read())
                     {
-                        Consumption consumption = new Consumption();
-                        consumption.Id = (int)reader["id"];
-                        consumption.FoodName = reader["food_name"].ToString();
-                        consumption.KCal = (float)Convert.ToDouble(reader["kcal"]);
-                        consumption.Protein = (float)Convert.ToDouble(reader["protein"]);
-                        consumption.Fiber = (float)Convert.ToDouble(reader["fiber"]);
-                        consumption.Calium = (float)Convert.ToDouble(reader["calium"]);
-                        consumption.Sodium = (float)Convert.ToDouble(reader["sodium"]);
-                        consumption.Amount = (int)reader["amount"];
-                        consumption.WeightUnitId = (int)reader["weight_unit_id"];
-                        consumption.Date = Convert.ToDateTime(reader["date"]).Date;
-                        consumption.PatientId = (int)reader["patient_id"];
-                        consumptions.Add(new ConsumptionView(consumption));
+                        consumption = new PatientConsumptionView();
+                        
+                        try
+                        {
+                            consumption.ConsumptionId = (int)reader["id"];
+                            consumption.FoodName = reader["food_name"].ToString();
+                            consumption.KCal = (float)Convert.ToDouble(reader["kcal"]);
+                            consumption.Protein = (float)Convert.ToDouble(reader["protein"]);
+                            consumption.Fiber = (float)Convert.ToDouble(reader["fiber"]);
+                            consumption.Calium = (float)Convert.ToDouble(reader["calium"]);
+                            consumption.Sodium = (float)Convert.ToDouble(reader["sodium"]);
+                            consumption.Amount = (int)reader["amount"];
+                            consumption.Weight = new WeightUnitModel
+                            {
+                                Id = (int)reader["weight_unit_id"],
+                                Short = (string)reader["short"],
+                                Unit = (string)reader["unit"]
+                            };
+                            consumption.Date = Convert.ToDateTime(reader["date"]).Date;
+                            consumption.Valid = true;
+                        }
+                        catch (Exception ex) {
+                            if (consumption.ConsumptionId == 0)
+                            {
+                                throw ex;
+                            }
+                        }
+                        
+                        // Adds incomplete consumption to consumptions. Consumption is Invalid by default
+                        consumptions.Add(consumption);
                     }
                 }
                 catch (Exception e)
                 {
-                    consumptions.Add(new ConsumptionView());
+                    conn.Close();
                     throw e;
                 }
             }
@@ -159,7 +158,7 @@ namespace AppNiZiAPI
             return consumptions;
         }
 
-        public bool UpdateConsumption(int consumptionId, ConsumptionView consumption, int patientId)
+        public bool UpdateConsumption(int consumptionId, Consumption consumption)
         {
             bool updated;
             var updateQuery = $"UPDATE Consumption SET " +
@@ -170,7 +169,7 @@ namespace AppNiZiAPI
             conn.Open();
             try
             {
-                updated = ConsumptionCommand(updateQuery, consumption, patientId).ExecuteNonQuery() > 0;
+                updated = ConsumptionCommand1(updateQuery, consumption).ExecuteNonQuery() > 0;
             }
             catch (Exception)
             {
@@ -180,24 +179,20 @@ namespace AppNiZiAPI
             return updated;
         }
 
-        private SqlCommand ConsumptionCommand(String query, Consumption consumption)
-        {
-            return ConsumptionCommand(query, new ConsumptionView(consumption), consumption.PatientId);
-        }
-
-        private SqlCommand ConsumptionCommand(String query, ConsumptionView consumptionView, int PatientId)
+        // TODO: Remove patientId and check earlier if patientId equals consumption.patient_id
+        private SqlCommand ConsumptionCommand1(string query, Consumption consumption)
         {
             SqlCommand command = new SqlCommand(query, conn);
-            command.Parameters.AddWithValue("@food_name", consumptionView.FoodName);
-            command.Parameters.AddWithValue("@kcal", consumptionView.KCal);
-            command.Parameters.AddWithValue("@protein", consumptionView.Protein);
-            command.Parameters.AddWithValue("@fiber", consumptionView.Fiber);
-            command.Parameters.AddWithValue("@calium", consumptionView.Calium);
-            command.Parameters.AddWithValue("@sodium", consumptionView.Sodium);
-            command.Parameters.AddWithValue("@amount", consumptionView.Amount);
-            command.Parameters.AddWithValue("@weight_unit_id", consumptionView.WeightUnitId);
-            command.Parameters.AddWithValue("@date", consumptionView.Date);
-            command.Parameters.AddWithValue("@patient_id", PatientId);
+            command.Parameters.AddWithValue("@food_name", consumption.FoodName);
+            command.Parameters.AddWithValue("@kcal", consumption.KCal);
+            command.Parameters.AddWithValue("@protein", consumption.Protein);
+            command.Parameters.AddWithValue("@fiber", consumption.Fiber);
+            command.Parameters.AddWithValue("@calium", consumption.Calium);
+            command.Parameters.AddWithValue("@sodium", consumption.Sodium);
+            command.Parameters.AddWithValue("@amount", consumption.Amount);
+            command.Parameters.AddWithValue("@weight_unit_id", consumption.WeightUnitId);
+            command.Parameters.AddWithValue("@date", consumption.Date);
+            command.Parameters.AddWithValue("@patient_id", consumption.PatientId);
             return command;
         }
